Persisting interactions
We start storing more details about sender->receiver communication:
 messages, message lengths, receiver outputs.
As a result, we would be able to analyse the emergent protocol by
passing inputs to the game and running post-analysis within the Callback
interface. While before, the only way for doing that would be to
re-implement agent interaction in a separate entity.